Added functionality to show captions on "CC" button mousemove.
When auto-show/auto-hide is disabled, the captions will be shown when
the user will move the mouse over the "CC" button. They will then
auto-hide after a while if the user doesn't continue to move the
mouse over the "CC" button, or over the captions themselves, or use
the keyboard to select a specific caption. If the mouse pointer is not
over the captions or the "CC" button, they will hide after a while.

This is the key in the fix for bug BLD-355: Transcript hovers over
over videoplayer itself for PKU videos.